+++ apps/rosen-service/src/constants.ts
@@ -3,6 +3,7 @@ export const SCANNER_API_TIMEOUT = 10_000;
 export const ERGO_SCANNER_INTERVAL = 2 * 60 * 1000;
 export const CARDANO_SCANNER_INTERVAL = 30 * 1000;
 export const BITCOIN_SCANNER_INTERVAL = 10 * 60 * 1000;
+export const DOGE_SCANNER_INTERVAL = 60 * 1000;
 export const ETHEREUM_SCANNER_INTERVAL = 60 * 1000;
 export const BINANCE_SCANNER_INTERVAL = 10 * 1000;
 export const ASSET_CALCULATOR_INTERVAL = 30 * 1000;
@@ -10,5 +11,9 @@ export const ASSET_CALCULATOR_INTERVAL = 30 * 1000;
 export const ERGO_SCANNER_LOGGER_NAME = 'ergo-scanner';
 export const CARDANO_SCANNER_LOGGER_NAME = 'cardano-scanner';
 export const BITCOIN_SCANNER_LOGGER_NAME = 'bitcoin-scanner';
+export const DOGE_SCANNER_LOGGER_NAME = 'doge-scanner';
 export const ETHEREUM_SCANNER_LOGGER_NAME = 'ethereum-scanner';
 export const BINANCE_SCANNER_LOGGER_NAME = 'binance-scanner';
+
+export const ETHEREUM_BLOCK_TIME = 12;
+export const BINANCE_BLOCK_TIME = 3;