+++ apps/rosen/app/_wallets/index.ts
@@ -1,6 +1,7 @@
 import { eternl } from './eternl';
 import { lace } from './lace';
 import { metamask } from './metamask';
+import { myDoge } from './my-doge';
 import { nautilus } from './nautilus';
 import { okx } from './okx';
 
@@ -10,4 +11,5 @@ export const wallets = {
   [metamask.name]: metamask,
   [nautilus.name]: nautilus,
   [okx.name]: okx,
+  [myDoge.name]: myDoge,
 } as const;