+++ wallets/okx/src/wallet.ts
@@ -5,7 +5,6 @@ import {
   DisconnectionFailedError,
   AddressRetrievalError,
   ConnectionRejectedError,
-  UnavailableApiError,
   UserDeniedTransactionSignatureError,
   Wallet,
   WalletTransferParams,
@@ -13,7 +12,7 @@ import {
 
 import { WalletConfig } from './types';
 
-export class OKXWallet implements Wallet {
+export class OKXWallet extends Wallet {
   icon = OKXIcon;
 
   name = 'OKX';
@@ -24,31 +23,33 @@ export class OKXWallet implements Wallet {
 
   supportedChains = [NETWORKS.bitcoin.key];
 
+  constructor(private config: WalletConfig) {
+    super();
+  }
+
   private get api() {
     return window.okxwallet.bitcoin;
   }
 
-  constructor(private config: WalletConfig) {}
-
-  async connect(): Promise<void> {
+  connect = async (): Promise<void> => {
     this.requireAvailable();
     try {
       await this.api.connect();
     } catch (error) {
       throw new ConnectionRejectedError(this.name, error);
     }
-  }
+  };
 
-  async disconnect(): Promise<void> {
+  disconnect = async (): Promise<void> => {
     this.requireAvailable();
     try {
       await this.api.disconnect();
     } catch (error) {
       throw new DisconnectionFailedError(this.name, error);
     }
-  }
+  };
 
-  async getAddress(): Promise<string> {
+  getAddress = async (): Promise<string> => {
     this.requireAvailable();
     const accounts = await this.api.getAccounts();
 
@@ -57,9 +58,9 @@ export class OKXWallet implements Wallet {
     if (!account) throw new AddressRetrievalError(this.name);
 
     return account;
-  }
+  };
 
-  async getBalance(token: RosenChainToken): Promise<bigint> {
+  getBalance = async (token: RosenChainToken): Promise<bigint> => {
     this.requireAvailable();
     const amount = await this.api.getBalance();
 
@@ -74,23 +75,19 @@ export class OKXWallet implements Wallet {
     ).amount;
 
     return wrappedAmount;
-  }
+  };
 
-  isAvailable(): boolean {
+  isAvailable = (): boolean => {
     return (
       typeof window.okxwallet !== 'undefined' && !!window.okxwallet.bitcoin
     );
-  }
+  };
 
-  requireAvailable() {
-    if (!this.isAvailable()) throw new UnavailableApiError(this.name);
-  }
-
-  async isConnected(): Promise<boolean> {
+  isConnected = async (): Promise<boolean> => {
     return !!window.okxwallet.selectedAddress;
-  }
+  };
 
-  async transfer(params: WalletTransferParams): Promise<string> {
+  transfer = async (params: WalletTransferParams): Promise<string> => {
     this.requireAvailable();
     const userAddress = await this.getAddress();
 
@@ -128,5 +125,5 @@ export class OKXWallet implements Wallet {
     const txId = await this.config.submitTransaction(signedPsbtHex, 'hex');
 
     return txId;
-  }
+  };
 }