+++ packages/ui-kit/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @rosen-bridge/ui-kit
 
+## 1.10.1
+
+### Patch Changes
+
+- Reducing font size for 'Pending...' in the Amount component ensures smoother rendering and a cleaner UI.
+- Updated dependencies
+  - @rosen-ui/constants@0.3.0
+  - @rosen-ui/utils@0.5.0
+
+## 1.10.0
+
+### Minor Changes
+
+- Add a 'More' button to display additional details in the snackbar context
+
+### Patch Changes
+
+- Improve the styling and adjust font sizes in the Amount component for a better appearance.
+- Updated @rosen-bridge/icons@1.3.0 dependency
+
 ## 1.9.2
 
 ### Patch Changes